+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 ### Enhancements
 
   * [Indentation] Indent listing inside square brackets. (#160)
-  * [Indentation] Ident correct after a binary sequence `<<1,2,3,4>>`.
+  * [Indentation] Indent of binary sequence inside match block
+  * [Indentation] Indent correct after a binary sequence `<<1,2,3,4>>`.
   * [Indentation] Indent correct after oneline `def ... do:` function
   * [Indentation] Correct behavior after last line in buffer. (#145)