+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## v2.2.0-dev (unreleased)
+## v2.2.0 2014/12/31
 
 ### Enhancements